feat：修改代码

+++ digital-modules/digital-human/src/main/java/com/flash/human/domain/DhAvatar.java
@@ -63,9 +63,9 @@ public class DhAvatar extends TenantEntity {
     private String modelPath;
 
     /**
-     * 状态（1-可用 2-禁用）
+     * 状态（0-可用 1-禁用）
      */
-    private Long status;
+    private Integer status;
 
 
 }
