+++ digital-modules/digital-human/src/main/java/com/flash/human/domain/DhAvatarCategory.java
@@ -38,9 +38,9 @@ public class DhAvatarCategory extends TenantEntity {
     private String name;
 
     /**
-     * 状态（1-可用 2-禁用）
+     * 状态（0-可用 1-禁用）
      */
-    private Long status;
+    private Integer status;
 
     /**
      * 备注
